requirements.txt
# Core LLM and Agent Framework
# Pinned to 0.3.x versions
langchain==0.3.27
langchain-core==0.3.76
langchain-anthropic==0.3.17
langsmith>=0.4.5
anthropic>=0.57.1

# OpenSearch Integration
opensearch-py>=2.4.0

# Web Framework
fastapi>=0.116.1
uvicorn[standard]>=0.35.0

# Data Validation and Serialization
pydantic>=2.10.6
pydantic-settings>=2.8.1

# Caching and Session Management
redis>=6.2.0

# Authentication and Security
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20

# Observability and Monitoring
opentelemetry-api>=1.35.0
opentelemetry-sdk>=1.35.0
opentelemetry-exporter-prometheus>=0.56b0
prometheus-client>=0.22.1

# Logging
structlog>=25.4.0
colorama>=0.4.6

# Streamlit UI
streamlit>=1.46.1

# HTTP Client
httpx>=0.28.1
aiofiles>=24.1.0

# Configuration Management
python-dotenv>=1.1.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests~=2.32.4
aiohttp~=3.12.14